feat(dir): allow file creation customization
This is required to allow for creation of directories, where files are
not "regular files", e.g. to implement networking

Unfortunately, it does not seem to be possible to use generics for this
due to downcasting used within the subcrate. This is an issue in
particular when a parent or child `Node` is downcasted to `Directory`

Signed-off-by: Roman Volosatovs <roman@profian.com>